+++ .gitignore
@@ -1,3 +1,2 @@
 gameoflife
-life.json
-data.life
+data.db
