+++ go.mod
@@ -1,5 +1,10 @@
-module gameoflife
+module gameoflifemodel
 
-go 1.15
+go 1.16
 
-require github.com/spf13/pflag v1.0.5
+require (
+	github.com/TwinProduction/go-color v1.0.0
+	github.com/spf13/pflag v1.0.5
+	gorm.io/driver/sqlite v1.1.4
+	gorm.io/gorm v1.21.6
+)
