+++ main.go
@@ -1,25 +1,21 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
+	finder "gameoflifemodel/finder"
+	lifeModel "gameoflifemodel/model"
+	"github.com/TwinProduction/go-color"
 	"github.com/spf13/pflag"
-	"io/ioutil"
 	"math/rand"
 	"strconv"
 	"strings"
 )
 
-type dataModel struct {
-	D    int
-	SIZE int
-	DATA string
-	B    string
-	S    string
-}
-
-func random() bool {
-	return rand.Uint64()&(1<<63) == 0
+func random(probability int) bool {
+	min := 0
+	max := 100
+	value := rand.Intn(max-min) + min
+	return value <= probability
 }
 
 func readRule(rule string) []int {
@@ -49,25 +45,21 @@ func readRule(rule string) []int {
 	return answer
 }
 
-func main() {
+func RunProgram(m lifeModel.MODEL) {
 	var (
 		showHelp        bool
-		finderMod       bool
-		outputFile      string
-		inputFile       string
 		dimension       int
 		size            int
+		attempts        int
 		B               string
 		S               string
 		countGeneration int
 		b               []int
 		s               []int
 		model3d         bool
-		details         bool
+		probability     int
+		fileName        string
 	)
-	pflag.BoolVarP(&finderMod, "find", "f", false, "find mod")
-	pflag.StringVarP(&outputFile, "out", "o", "data.life", "output file")
-	pflag.StringVarP(&inputFile, "in", "i", "", "input file")
 	pflag.IntVarP(&dimension, "dimension", "d", 3, "dimension of world")
 	pflag.IntVarP(&size, "size", "S", 128, "side size")
 	pflag.StringVarP(&B, "b-rule", "b", "5", "Rules for birth")
@@ -75,7 +67,9 @@ func main() {
 	pflag.IntVarP(&countGeneration, "count", "g", 100, "count generations.")
 	pflag.BoolVarP(&showHelp, "help", "h", false, "Show help message")
 	pflag.BoolVarP(&model3d, "model3d", "3", false, "Use 3D model")
-	pflag.BoolVarP(&details, "details", "e", false, "save all generations")
+	pflag.IntVarP(&attempts, "attempt", "a", 100, "Count attempts")
+	pflag.IntVarP(&probability, "probability", "p", 50, "probability in %")
+	pflag.StringVarP(&fileName, "out", "o", "output.db", "Database name")
 	pflag.Parse()
 	if showHelp {
 		pflag.Usage()
@@ -84,76 +78,37 @@ func main() {
 		return
 	}
 
-	var model MODEL
+	var model lifeModel.MODEL
 
-	if model3d {
-		model = &Life3d{
-			SIZE: size,
-			N:    3,
-		}
-		dimension = 3
-	} else {
-		model = &Life{
-			SIZE: size,
-			N:    dimension,
-		}
-	}
-	var d []bool
-	if inputFile == "" {
-		d = make([]bool, intPow(size, dimension))
-		for i := range d {
-			d[i] = random() && random()
-		}
-	} else {
-		byteData, err := ioutil.ReadFile(inputFile)
-		if err != nil {
-			panic(err)
-		}
-		var md dataModel
-		json.Unmarshal(byteData, &md)
-		B = md.B
-		S = md.S
-		model.setN(md.D)
-		model.setSIZE(md.SIZE)
-		str_data := RLEDecode(md.DATA)
-		for _, c := range str_data {
-			if string(c) == "A" {
-				d = append(d, true)
-			} else {
-				d = append(d, false)
+	if m == nil {
+		if model3d {
+			model = &lifeModel.Life3d{
+				SIZE: size,
+				N:    3,
+			}
+			dimension = 3
+		} else {
+			model = &lifeModel.Life{
+				SIZE: size,
+				N:    dimension,
 			}
 		}
+	} else {
+		model = m
+	}
+	d := make([]bool, lifeModel.IntPow(size, dimension))
+	for i := range d {
+		d[i] = random(probability)
 	}
 	b = readRule(B)
 	s = readRule(S)
 
 	model.Setup(b, s, d) // Set rules and data, if data exists
-	fmt.Println("HERE")
-	fmt.Println(model.GetData())
-	fmt.Println("Model is created")
-	if finderMod {
-		for t := 0; t < 100; t++ {
-			findRules(model, countGeneration, details, t)
-		}
-		return
-	}
-	for i := countGeneration; i > 0; i-- {
-		model.NextGeneration()
-	}
-	out := RLECode(dataToString(model.GetData()))
-	fmt.Println("Starting write to file")
-	output := dataModel{
-		D:    model.getN(),
-		SIZE: model.getSIZE(),
-		DATA: out,
-		B:    B,
-		S:    S,
-	}
-	file, _ := json.MarshalIndent(output, "", "")
-	saveErr := ioutil.WriteFile(outputFile, file, 0644)
-	if saveErr != nil {
-		panic(saveErr)
-	} else {
-		fmt.Println("Finish write to file. No errors")
-	}
+	fmt.Println(color.Ize(color.Green, "Start game of life"))
+	finder.FindRules(model, countGeneration, attempts, fileName)
+	fmt.Println(color.Ize(color.Green, "Finish. No Errors"))
+}
+
+func main() {
+	RunProgram(nil)
 }
